server/requirements.txt
Flask 
flask-cors 
python-dotenv 
supabase 
requests
py-cord
python-dotenv